Initial commit. It works, but it has a race condition.
See line 28, the call to time.sleep(). There is no guarantee that the
user will receive OP from Chanserv before that time.

Strings are currently hardcoded. In the current state it only works
for the #gultij channel on Freenode (due to its dependency on Chanserv
to provide OP, and the way it requests it).